Fixed getGraphQLQuery and reorganized classes in packages
